requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.7
PyJWT==1.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
social-auth-app-django==2.1.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
dj-database-url==0.4.0
psycopg2==2.7.4
braintree==3.39.1
django-ckeditor==5.4.0